Move embed handler, fix linked DM, fix embed color
